Apply glow filter in Renderer
