+++ src/Video/Renderer.cpp
@@ -8,6 +8,8 @@
 #include "PostProcessing/PostProcessing.hpp"
 #include "PostProcessing/ColorFilter.hpp"
 #include "PostProcessing/GammaCorrectionFilter.hpp"
+#include "PostProcessing/GlowBlurFilter.hpp"
+#include "PostProcessing/GlowFilter.hpp"
 #include "RenderTarget.hpp"
 
 using namespace Video;
@@ -20,6 +22,8 @@ Renderer::Renderer(const glm::vec2& screenSize) {
     postProcessing = new PostProcessing(screenSize);
     colorFilter = new ColorFilter(glm::vec3(1.f, 1.f, 1.f));
     gammaCorrectionFilter = new GammaCorrectionFilter();
+    glowFilter = new GlowFilter();
+    glowBlurFilter = new GlowBlurFilter();
 }
 
 Renderer::~Renderer() {
@@ -73,6 +77,17 @@ void Renderer::PrepareSkinnedMeshRendering(const glm::mat4& viewMatrix, const gl
 
 void Renderer::RenderSkinnedMesh(const Video::Geometry::Geometry3D* geometry, const Video::Texture* diffuseTexture, const Video::Texture* normalTexture, const Video::Texture* specularTexture, const Video::Texture* glowTexture, const glm::mat4& modelMatrix, const std::vector<glm::mat4>& bones, const std::vector<glm::mat3>& bonesIT) {
     skinRenderProgram->Render(geometry, diffuseTexture, normalTexture, specularTexture, glowTexture, modelMatrix, bones, bonesIT);
+}
+
+void Renderer::ApplyGlow(int blurAmount) {
+    glowBlurFilter->SetScreenSize(screenSize);
+    for (int i = 0; i < blurAmount; ++i) {
+        glowBlurFilter->SetHorizontal(true);
+        postProcessing->ApplyFilter(glowBlurFilter);
+        glowBlurFilter->SetHorizontal(false);
+        postProcessing->ApplyFilter(glowBlurFilter);
+    }
+    postProcessing->ApplyFilter(glowFilter);
 }
 
 void Renderer::ApplyColorFilter(const glm::vec3& color) {
